Move list_packages command to separate module
